+++ app/src/main/res/layout/activity_main.xml
@@ -16,4 +16,33 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         />
+    <Button
+        style="@style/button"
+        android:id="@+id/Likes"
+        android:text="@string/favorites"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.8"
+        />
+
+    <Button
+    style="@style/button"
+    android:id="@+id/addPoster"
+    android:text="@string/favorites"
+    app:layout_constraintBottom_toBottomOf="parent"
+    app:layout_constraintStart_toStartOf="parent"
+    app:layout_constraintEnd_toEndOf="parent"
+    app:layout_constraintHorizontal_bias="0.8"/>
+
+    <Button
+        style="@style/button"
+        android:id="@+id/delPoster"
+        android:text="@string/favorites"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.8"/>
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
