+++ app/src/main/res/layout/poster.xml
@@ -42,8 +42,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         android:layout_width="24dp"
         android:layout_height="24dp"
-
-        android:onClick="onClickLike"/>
+        android:onClick="onLike"/>
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
